[ai] Extract `parseGitDiffText` in auto-code-review GitCLI to a new file
In the CI runs of auto-code-review (KT-87721), the diffs need to be
obtained not from the Git CLI, but from GitHub. The diff format is
exactly the same, though.

So this commit extracts the function for parsing that format,
`parseGitDiffText`, to a separate file to allow reusing it.
